auth-service/requirements.txt
Flask==2.0.1
SQLAlchemy==1.4.22
Werkzeug==2.0.1
python-dotenv==0.19.2